+++ app/src/main/java/co/com/bankuish/mobile/jaime/challenge/infrastructure/data_source/DataSource.kt
@@ -1,6 +1,6 @@
 package co.com.bankuish.mobile.jaime.challenge.infrastructure.data_source
 
-import co.com.bankuish.mobile.jaime.challenge.infrastructure.model.RepositoriesGitHubDTO
+import co.com.bankuish.mobile.jaime.challenge.infrastructure.dtos.RepositoriesGitHubDTO
 import retrofit2.Response
 
 internal interface DataSource {
