+++ app/src/main/java/co/com/bankuish/mobile/jaime/challenge/infrastructure/models/Author.kt
@@ -1,4 +1,4 @@
-package co.com.bankuish.mobile.jaime.challenge.infrastructure.model
+package co.com.bankuish.mobile.jaime.challenge.infrastructure.models
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
